+++ src/sunmoon.h
@@ -1,1 +1,1 @@
-void sunmooncalc(double jd, int tz, float lat, float lon, int iobj, float* utrise, float* utset);
+void sunmooncalc(double jd, float tz, float lat, float lon, int iobj, float* utrise, float* utset);
